# docs/swagger.yaml
definitions:
  errcode.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.NameValid:
    properties:
      isValid:
        type: boolean
    type: object
  github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.ResetPwd:
    properties:
      newPassword:
        type: string
      originPassword:
        type: string
      userName:
        type: string
    type: object
  github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.UserItem:
    properties:
      name:
        type: string
      spec:
        $ref: '#/definitions/v1.UserSpec'
      status:
        $ref: '#/definitions/v1.UserStatus'
    type: object
  github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.UserList:
    properties:
      items:
        items:
          $ref: '#/definitions/github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.UserItem'
        type: array
      total:
        type: integer
    type: object
...
  response.SuccessInfo:
    properties:
      message:
        type: string
    type: object
...
  v1.UserSpec:
    properties:
      displayName:
        type: string
      email:
        type: string
      language:
        description: 'The preferred written or spoken language for the user: chinese/english'
        type: string
      loginType:
        description: Login method used, normal/openId/ldap
        type: string
      password:
        type: string
      phone:
        type: string
      state:
        type: string
    type: object
  v1.UserStatus:
    properties:
      lastLoginIP:
        type: string
      lastLoginTime:
        description: The user status, normal/forbidden
        type: string
    type: object
...
info:
  contact: {}
  description: This is KubeWorkz api documentation.
  title: Swagger KubeWorkz API
  version: "1.0"
paths:
...
  /api/v1/kube/user:
    get:
      description: fuzzy query user by name or displayName
      parameters:
      - description: keyword for query
        in: query
        name: query
        type: string
      - description: page size
        in: query
        name: pageSize
        type: integer
      - description: page num
        in: query
        name: pageNum
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.UserList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: list user
      tags:
      - user
    post:
      consumes:
      - application/json
      description: create user manually
      parameters:
      - description: user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/v1.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: create user
      tags:
      - user
  /api/v1/kube/user/:username:
    get:
      description: check username when update user password
      parameters:
      - description: user name
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.NameValid'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: check username
      tags:
      - user
    put:
      consumes:
      - application/json
      description: update user information
      parameters:
      - description: user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/v1.User'
      - description: user name
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: update user
      tags:
      - user
  /api/v1/kube/user/kubeconfigs:
    get:
      description: fetch kubeConfig for specified user
      parameters:
      - description: specified user for kubeconfig
        in: query
        name: user
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: fetch kubeConfigs
      tags:
      - user
  /api/v1/kube/user/pwd:
    put:
      description: update user password
      parameters:
      - description: user old and new password
        in: body
        name: resetPwd
        required: true
        schema:
          $ref: '#/definitions/github.com_saashqdev_kubeworkz_pkg_apiserver_kubeapi_user.ResetPwd'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: update password
      tags:
      - user
...
  /api/v1/kube/user/users:
    post:
      consumes:
      - multipart/form-data
      description: import and create users from CSV file in batches
      parameters:
      - description: file
        in: formData
        name: userInfoFile
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errcode.ErrorInfo'
      summary: import user
      tags:
      - user
swagger: "2.0"